typed cast to double for getValue

+++ RationalNumber.java
@@ -26,7 +26,7 @@ public class RationalNumber extends RealNumber
   }
 
   public double getValue(){
-    return (numerator/denominator);
+    return (double)numerator/denominator;
   }
 
   /**
